+++ src/leiningen/spec.clj
@@ -1,12 +1,8 @@
 (ns leiningen.spec
   (:use
-    [speclj.running :only (*runner* run-and-report)]
-    [speclj.run.standard :only (new-standard-runner)]
-    [speclj.reporting :only (active-reporter)]
-    [leiningen.compile :only [eval-in-project]])
-  (:import
-    [java.io File]))
+    [leiningen.compile :only [eval-in-project]]))
 
 (defn spec [project & args]
-  (let [exec-form `(binding [~'invoke-method "lein spec"] (~'run ~@args))]
+  (let [args (cons "-c" args)
+        exec-form `(binding [~'invoke-method "lein spec"] (~'run ~@args))]
     (eval-in-project project exec-form nil nil `(use 'speclj.main))))